Added box model padding and margin

+++ DMXUI_Example/AppliedGraphicsProgramming/WindowHandler.h
@@ -13,8 +13,8 @@ public:
 	static LRESULT CALLBACK WinProc(_In_ HWND aHWND, _In_ UINT aUMSg, _In_ WPARAM aWParam, _In_ LPARAM aLParam);
 	bool Init(const SWindowData& aWindowData);
 	HWND& GetWindowHandle();
-	__forceinline float GetWidth() { return myWindowData.myWidth; }
-	__forceinline float GetHeight() {return myWindowData.myHeight; }
+	__forceinline float GetWidth() { return static_cast<float>(myWindowData.myWidth ); }
+	__forceinline float GetHeight() {return static_cast<float>(myWindowData.myHeight); }
 private:
 	HWND myWindowHandle;
 };
